+++ src/lib/iTin.Utilities/iTin.Utilities.Pdf/iTin.Utilities.Pdf.Writer/ComponentModel/Operations/Items/Insert/Base/InsertBase.cs
@@ -5,6 +5,7 @@ using System.IO;
 using iTin.Core.Helpers;
 using iTin.Core.Models.Design.Styling;
 
+using iTin.Utilities.Pdf.Writer.ComponentModel.Input;
 using iTin.Utilities.Pdf.Writer.ComponentModel.Result.Insert;
 
 namespace iTin.Utilities.Pdf.Writer.ComponentModel