+++ src/lib/iTin.Utilities/iTin.Utilities.Pdf/iTin.Utilities.Pdf.Writer/ComponentModel/Operations/Items/Set/Base/SetBase.cs
@@ -1,90 +1,50 @@
 ﻿
-using System.IO;
+using System;
+using System.ComponentModel;
 
-using iTin.Core.Helpers;
-
-using iTin.Utilities.Pdf.Writer.ComponentModel.Result.Set;
+using iTin.Core.ComponentModel;
 
 namespace iTin.Utilities.Pdf.Writer.ComponentModel
 {
     /// <summary>
-    /// Specialization of <see cref="ISet"/> interface.<br/>
-    /// Acts as base class for set actions
+    /// A Specialization of <see cref="ISet"/> class.<br/>
     /// </summary>
-    /// <seealso cref="ISet"/>
-    public abstract class SetBase : ISet
+    public class SetBase : ISet
     {
-        #region interface
-
-        #region ISet
-
-        #region public methods
-
-        #region [public] (SetResult) Apply(string, IInput): Try to execute the set action 
         /// <summary>
-        /// Try to execute the set action.
+        /// 
         /// </summary>
-        /// <param name="file">file input</param>
-        /// <param name="context">Input context</param>
-        /// <returns>
-        /// <para>
-        /// A <see cref="SetResult"/> reference that contains the result of the operation, to check if the operation is correct, the <b>Success</b>
-        /// property will be <b>true</b> and the <b>Value</b> property will contain the value; Otherwise, the the <b>Success</b> property
-        /// will be false and the <b>Errors</b> property will contain the errors associated with the operation, if they have been filled in.
-        /// </para>
-        /// <para>
-        /// The type of the return value is <see cref="SetResultData"/>, which contains the operation result
-        /// </para>
-        /// </returns>
-        public SetResult Apply(string file, IInput context) => Apply(StreamHelper.TextFileToStream(file), context);
-        #endregion
+        protected SetBase()
+        {
+            Key = GetKeyDescription();
+        }
 
-        #region [public] (SetResult) Apply(Stream, IInput): Try to execute the set action 
         /// <summary>
-        /// Try to execute the set action.
+        /// 
         /// </summary>
-        /// <param name="input">stream input</param>
-        /// <param name="context">Input context</param>
-        /// <returns>
-        /// <para>
-        /// A <see cref="SetResult"/> reference that contains the result of the operation, to check if the operation is correct, the <b>Success</b>
-        /// property will be <b>true</b> and the <b>Value</b> property will contain the value; Otherwise, the the <b>Success</b> property
-        /// will be false and the <b>Errors</b> property will contain the errors associated with the operation, if they have been filled in.
-        /// </para>
-        /// <para>
-        /// The type of the return value is <see cref="SetResultData"/>, which contains the operation result
-        /// </para>
-        /// </returns>
-        public SetResult Apply(Stream input, IInput context) => SetImpl(input, context);
-        #endregion
+        public string Key { get; }
 
-        #endregion
-
-        #endregion
-
-        #endregion
-
-        #region protected abtract methods
-
-        #region [public] {abstract} (SetResult) SetImpl(Stream, IInput): Implementation to execute when set action 
         /// <summary>
-        /// Implementation to execute when set action.
+        /// 
         /// </summary>
-        /// <param name="input">stream input</param>
-        /// <param name="context">Input context</param>
-        /// <returns>
-        /// <para>
-        /// A <see cref="SetResult"/> reference that contains the result of the operation, to check if the operation is correct, the <b>Success</b>
-        /// property will be <b>true</b> and the <b>Value</b> property will contain the value; Otherwise, the the <b>Success</b> property
-        /// will be false and the <b>Errors</b> property will contain the errors associated with the operation, if they have been filled in.
-        /// </para>
-        /// <para>
-        /// The type of the return value is <see cref="SetResultData"/>, which contains the operation result
-        /// </para>
-        /// </returns>
-        protected abstract SetResult SetImpl(Stream input, IInput context);
-        #endregion
+        public string Value { get; set; }
 
-        #endregion
+
+        /// <summary>
+        /// Returns the value of attribute of type <see cref="T:iTin.Core.ComponentModel.KeyDescriptionAttribute"/> for this enum value. 
+        /// If this attribute is not defined returns <strong>null</strong> (<b>Nothing</b> in Visual Basic)
+        /// </summary>
+        /// <returns>
+        /// A <see cref="T:System.String"/> that contains the value of attribute.
+        /// </returns>
+        private string GetKeyDescription()
+        {
+            var type = GetType();
+            
+            DescriptionAttribute attr = Attribute.GetCustomAttribute(type, typeof(KeyDescriptionAttribute)) as KeyDescriptionAttribute;
+            var result = attr?.Description;
+
+            return result;
+        }
     }
 }