+++ src/lib/iTin.Utilities/iTin.Utilities.Pdf/iTin.Utilities.Pdf.Writer/ComponentModel/Operations/Results/Replace/ReplaceResult.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 
 using iTin.Core.ComponentModel;
 
+using iTin.Utilities.Pdf.Writer.ComponentModel.Input;
 using iTin.Utilities.Pdf.Writer.ComponentModel.Result.Insert;
 using iTin.Utilities.Pdf.Writer.ComponentModel.Result.Set;
 
@@ -133,40 +134,6 @@ namespace iTin.Utilities.Pdf.Writer.ComponentModel.Result.Replace
         }
         #endregion
 
-        #region [public] (SetResult) Replace(ISet, bool = true): Try to replace an element in this input
-        /// <summary>
-        /// Try to replace an element in this input.
-        /// </summary>
-        /// <param name="data">Reference to replaceable object information</param>
-        /// <param name="canSet">Determines if can set. Default is <b>true</b>.</param>
-        /// <returns>
-        /// Operation result.
-        /// </returns>
-        public SetResult Set(ISet data, bool canSet = true)
-        {
-            if (!canSet)
-            {
-                return data == null
-                    ? SetResult.CreateErroResult("Missing data")
-                    : SetResult.CreateSuccessResult(new SetResultData
-                    {
-                        Context = Result.Context,
-                        InputStream = Result.OutputStream,
-                        OutputStream = Result.OutputStream
-                    });
-            }
-
-            SetResult result = SetImplStrategy(data, Result.Context);
-
-            if (Result.Context.AutoUpdateChanges)
-            {
-                Result.Context.Input = result.Result.OutputStream;
-            }
-
-            return result;
-        }
-        #endregion
-
         #endregion
 
         #region private methods
@@ -177,8 +144,8 @@ namespace iTin.Utilities.Pdf.Writer.ComponentModel.Result.Replace
         //private ReplaceResult ReplaceImplStrategy(IReplace data, IInput context)
         //    => data == null ? ReplaceResult.CreateErroResult("Missing data") : data.Apply(Result.OutputStream, context);
 
-        private SetResult SetImplStrategy(ISet data, IInput context)
-            => data == null ? SetResult.CreateErroResult("Missing data") : data.Apply(Result.OutputStream, context);
+        //private SetResult SetImplStrategy(ISet data, IInput context)
+        //    => data == null ? SetResult.CreateErroResult("Missing data") : data.Apply(Result.OutputStream, context);
 
         #endregion
     }